Shift logic to SciPaMaToPubmedArticle using GreekLetterTranslator
