+++ core/core-logic/src/test/kotlin/ch/difty/scipamato/core/logic/parsing/AuthorParsingTest.kt
@@ -3,14 +3,10 @@
 package ch.difty.scipamato.core.logic.parsing
 
 import org.amshove.kluent.shouldBeEqualTo
-import org.amshove.kluent.shouldBeFalse
 import org.amshove.kluent.shouldBeInstanceOf
 import org.amshove.kluent.shouldBeNull
 import org.amshove.kluent.shouldContain
 import org.amshove.kluent.shouldContainSame
-import org.amshove.kluent.shouldNotBeEqualTo
-import org.amshove.kluent.shouldNotContain
-import org.amshove.kluent.shouldStartWith
 import org.junit.jupiter.api.Test
 
 internal class PubmedAuthorParserTest {
@@ -111,57 +107,6 @@ internal class PubmedAuthorParserTest {
         p = PubmedAuthorParser("Flückiger P, Bäni HU.")
         p.authors.map { it.lastName } shouldContainSame listOf("Flückiger", "Bäni")
         p.authors.map { it.firstName } shouldContainSame listOf("P", "HU")
-    }
-
-    @Test
-    fun canParsePmId35469927() {
-        val authors = "Markozannes G, Pantavou K, Rizos EC, Sindosi OΑ, Tagkas C, Seyfried M, " +
-            "Saldanha IJ, Hatzianastassiou N, Nikolopoulos GK, Ntzani E."
-        p = PubmedAuthorParser(authors)
-        p.authors.map { it.lastName } shouldContainSame listOf(
-            "Markozannes", "Pantavou", "Rizos", "Sindosi", "Tagkas", "Seyfried",
-            "Saldanha", "Hatzianastassiou", "Nikolopoulos", "Ntzani"
-        )
-    }
-
-    @Test
-    fun canParsePmId35469927_resulting_in_validA() {
-        val authorsStringPmId35469927 = "Markozannes G, Pantavou K, Rizos EC, Sindosi OΑ, Tagkas C, " +
-            "Seyfried M, Saldanha IJ, Hatzianastassiou N, Nikolopoulos GK, Ntzani E."
-
-        // Ensure the fourth author has a capital greek letter alpha in it (visually nearly identical to "A")
-        val fourthAuthor = authorsStringPmId35469927.split(", ")[3]
-        fourthAuthor shouldStartWith "Sindosi"
-        val capitalGreekAlpha = Char(913) // Capital Greek Letter Alpha
-        fourthAuthor shouldContain capitalGreekAlpha
-
-        p = PubmedAuthorParser(authorsStringPmId35469927)
-
-        p.authorsString shouldNotContain capitalGreekAlpha
-        p.authorsString shouldContain Char(65) // Regular Capital A
-    }
-
-    @Test
-    fun replacesSeveralGreekLetters_with_visuallySimilarLetters() {
-        val wGreeks = "ΑΒΕΖΗΙΚΜΝΟΡΤΥΧ Α."
-        val woGreek = "ABEZHIKMNOPTYX A."
-
-        assertLettersAreNotSame(wGreeks, woGreek)
-        @Suppress("KotlinConstantConditions")
-        (wGreeks == woGreek).shouldBeFalse()
-
-        p = PubmedAuthorParser(wGreeks)
-        p.authorsString shouldBeEqualTo woGreek
-
-    }
-
-    @Suppress("SameParameterValue")
-    private fun assertLettersAreNotSame(one: String, two: String) {
-        val toIgnore = setOf(' ', '.')
-        val first = one.filterNot { it in toIgnore }
-        val second = two.filterNot { it in toIgnore }
-        first.length shouldBeEqualTo second.length
-        first.forEachIndexed { index, char -> char shouldNotBeEqualTo second[index] }
     }
 }
