+++ core/core-pubmed-api/src/main/java/ch/difty/scipamato/core/pubmed/PubmedXmlService.java
@@ -108,7 +108,7 @@ public class PubmedXmlService implements PubmedArticleService {
      * <ul>
      * <li>via API, e.g.
      * {@code https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi?db=pubmed&id=25395026&retmode=xml}</li>
-     * <li>through the Web UI (e.g. https://www.ncbi.nlm.nih.gov/pubmed/25395026)
+     * <li>through the Web UI (e.g. <a href="https://www.ncbi.nlm.nih.gov/pubmed/25395026">...</a>)
      * when sending to {@code file} in format {@code XML}</li>
      * </ul>
      *
@@ -127,7 +127,7 @@ public class PubmedXmlService implements PubmedArticleService {
             articles.addAll(article
                 .stream()
                 .map(PubmedArticleFacade::newPubmedArticleFrom)
-                .collect(Collectors.toList()));
+                .toList());
         } catch (final Exception e) {
             log.info("Unable to parse xmlString '{}': {}", xmlString, e.getMessage());
         }
